back/requirements.txt
click==6.6
Flask==0.11.1
Flask-Cors==3.0.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Preston==2.0.3
PyJWT==1.4.2
requests==2.11.1
six==1.10.0
Werkzeug==0.11.11
xmltodict==0.10.2
